added hierarchical labeling and query functions
