+++ cytopus/__init__.py
@@ -1,4 +1,4 @@
 """A Knowledge Base for Single Cell Biology"""
-from . import tools as tl
-from .tools import *
+
+from .tl import *
 from .knowledge_base import KnowledgeBase, get_data
