--- cytopus/tools/__init__.py
@@ -1,3 +1,0 @@
-"""Tools to use KnowledgeBase to label and interpret data"""
-from .label import overlap_coefficient, label_marker_genes, get_celltype, get_gmt, flatten_hierarchical_dict, hierarchy_to_csv, geneset_to_csv, metadata_to_csv
-from .create import construct_kb
